proba 1. W Human trzeba było zmienić typ wartości age, bo się nie zgadzał z testem.

+++ include/Human.hpp
@@ -6,18 +6,18 @@ class Human
 {
 public:
     Human() : name{"Anonymous"}, age{0}, likes_dogs{true}, likes_cats{true} {}
-    Human(std::string name_, unsigned int age_, bool ld, bool lk)
+    Human(std::string name_, size_t age_, bool ld, bool lk)
         : name{std::move(name_)}, age{age_}, likes_dogs{ld}, likes_cats{lk}
     {}
 
     const std::string& getName() const { return name; }
-    unsigned int       getAge() const { return age; }
+    size_t              getAge() const { return age; }
     bool               isMonster() const { return !likes_dogs && !likes_cats; }
     void               birthday() { ++age; }
 
 private:
     std::string  name;
-    unsigned int age;
+    size_t age;
     bool         likes_dogs;
     bool         likes_cats;
 };
